move some tests to corresponding location
