+++ in_toto/resolver/_file_resolver.py
@@ -18,62 +18,59 @@ class FileResolver(Resolver):
   @classmethod
   def resolve_uri(cls, generic_uri, exclude_patterns=None,
                   follow_symlink_dirs=False):
-    norm_paths = []
 
-    for norm_path in super().apply_exclude_patterns(
-        [os.path.normpath(generic_uri)], exclude_patterns):
+    norm_paths = super().apply_exclude_patterns(
+        [os.path.normpath(generic_uri)], exclude_patterns)
+    if not norm_paths:
+      return norm_paths
+    norm_path = norm_paths.pop()
 
-      if os.path.isfile(norm_path):
-        norm_path = norm_path.replace('\\', '/')
-        norm_paths.append(norm_path)
+    if os.path.isfile(norm_path):
+      return [norm_path.replace('\\', '/')]
 
-      elif os.path.isdir(norm_path):
-        for root, dirs, files in os.walk(norm_path,
-                                         followlinks=follow_symlink_dirs):
-          dirpaths = []
-          for dirname in dirs:
-            npath = os.path.normpath(os.path.join(root, dirname))
-            dirpaths.append(npath)
+    if not os.path.isdir(norm_path):
+      LOG.info("path: {} does not exist, skipping..".format(norm_path))
+      return norm_paths
 
-          if exclude_patterns:
-            dirpaths = super().apply_exclude_patterns(dirpaths,
-                                                      exclude_patterns)
+    for root, dirs, files in os.walk(norm_path,
+        followlinks=follow_symlink_dirs):
 
-          dirs[:] = []
-          for dirpath in dirpaths:
-            name = os.path.basename(dirpath)
-            dirs.append(name)
+      # Create a list of normalized dirpaths
+      dirpaths = []
+      for dirname in dirs:
+        npath = os.path.normpath(os.path.join(root, dirname))
+        dirpaths.append(npath)
 
-          filepaths = []
-          for filename in files:
-            norm_filepath = os.path.normpath(os.path.join(root, filename))
+      if exclude_patterns:
+        dirpaths = super().apply_exclude_patterns(dirpaths, exclude_patterns)
 
-            if os.path.isfile(norm_filepath):
-              filepaths.append(norm_filepath)
+      dirs[:] = [os.path.basename(d) for d in dirpaths]
 
-            else:
-              LOG.info("File '{}' appears to be a broken symlink. Skipping..."
-                  .format(norm_filepath))
+      filepaths = []
+      for filename in files:
+        norm_filepath = os.path.normpath(os.path.join(root, filename))
 
-          if exclude_patterns:
-            filepaths = super().apply_exclude_patterns(filepaths,
-                                                exclude_patterns)
+        if os.path.isfile(norm_filepath):
+          filepaths.append(norm_filepath)
 
-          for filepath in filepaths:
-            normalized_filepath = filepath.replace("\\", "/")
-            norm_paths.append(normalized_filepath)
+        else:
+          LOG.info("File '{}' appears to be a broken symlink. Skipping..."
+              .format(norm_filepath))
 
-      else:
-        LOG.info("path: {} does not exist, skipping..".format(norm_path))
+      if exclude_patterns:
+        filepaths = super().apply_exclude_patterns(filepaths, exclude_patterns)
+
+      for filepath in filepaths:
+        normalized_filepath = filepath.replace("\\", "/")
+        norm_paths.append(normalized_filepath)
 
     return norm_paths
 
   @classmethod
   def get_hashable_representation(cls, resolved_uri,
                                   normalize_line_endings=False):
-    """Internal helper that takes a filename and hashes the respective file's
-    contents using the passed hash_algorithms and returns a hashdict conformant
-    with securesystemslib.formats.HASHDICT_SCHEMA. """
+    """Internal helper that takes a filename and creates a hashable
+    representation of the file contents."""
 
     data = b""
