+++ tests/test_runlib.py
@@ -34,8 +34,8 @@ import in_toto.exceptions
 from in_toto.models.metadata import Envelope, Metablock
 from in_toto.exceptions import SignatureVerificationError
 from in_toto.runlib import (in_toto_run, in_toto_record_start,
-    in_toto_record_stop, record_artifacts_as_dict, _apply_exclude_patterns,
-    _hash_artifact, _subprocess_run_duplicate_streams)
+    in_toto_record_stop, record_artifacts_as_dict, _hash_artifact,
+    _subprocess_run_duplicate_streams)
 from securesystemslib.interface import (
     generate_and_write_unencrypted_rsa_keypair,
     import_rsa_privatekey_from_file,
@@ -47,52 +47,6 @@ import securesystemslib.exceptions
 
 from tests.common import TmpDirMixin
 from pathlib import Path
-
-
-class Test_ApplyExcludePatterns(unittest.TestCase):
-  """Test _apply_exclude_patterns(names, exclude_patterns) """
-
-  def test_apply_exclude_explict(self):
-    names = ["foo", "bar", "baz"]
-    patterns = ["foo", "bar"]
-    expected = ["baz"]
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(sorted(result), sorted(expected))
-
-  def test_apply_exclude_all(self):
-    names = ["foo", "bar", "baz"]
-    patterns = ["*"]
-    expected = []
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(sorted(result), sorted(expected))
-
-  def test_apply_exclude_multiple_star(self):
-    names = ["foo", "bar", "baz"]
-    patterns = ["*a*"]
-    expected = ["foo"]
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(result, expected)
-
-  def test_apply_exclude_question_mark(self):
-    names = ["foo", "bazfoo", "barfoo"]
-    patterns = ["ba?foo"]
-    expected = ["foo"]
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(result, expected)
-
-  def test_apply_exclude_seq(self):
-    names = ["baxfoo", "bazfoo", "barfoo"]
-    patterns = ["ba[xz]foo"]
-    expected = ["barfoo"]
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(result, expected)
-
-  def test_apply_exclude_neg_seq(self):
-    names = ["baxfoo", "bazfoo", "barfoo"]
-    patterns = ["ba[!r]foo"]
-    expected = ["barfoo"]
-    result = _apply_exclude_patterns(names, patterns)
-    self.assertListEqual(result, expected)
 
 
 class TestRecordArtifactsAsDict(unittest.TestCase, TmpDirMixin):
